chore(refactor): Merge all errors back into single error module
